+++ BTCPayServer/Hosting/BTCPayServerServices.cs
@@ -357,7 +357,7 @@ namespace BTCPayServer.Hosting
 
             services.AddReportProvider<PaymentsReportProvider>();
             services.AddReportProvider<OnChainWalletReportProvider>();
-            services.AddReportProvider<ItemsReportProvider>();
+            services.AddReportProvider<ProductsReportProvider>();
 
             services.AddHttpClient(WebhookSender.OnionNamedClient)
                 .ConfigurePrimaryHttpMessageHandler<Socks5HttpClientHandler>();
